* iris-mp-dev: (31 commits)
  Fixes in user kylas id update & user access token update & other fixes
  [IRISMP-113] - changes added
  Multiple Fixes encountered while running the entire flow from new client sign up. - Fixed kylas user id getting nil & put required validations
  [IRISMP-113] - Scheme changes added
  [IRISMP-113] - Sales Head - Booking changes added
  [IRISMP-113] - send payment link label changes added
  [IRISMP-113] - sales head - Lead Index page
  [IRISMP-113] - sales head - Lead details pages changes added
  Fix policies for exports & project links
  Show KYC list page in marketplace
  [IRISMP-113] - Hide tentative agreement date
  [IRISMP-112] - Label for oauth type
  [IRISMP-112] - Created a sync lead worker that handles pushing deal and contact to kylas - Uniqueness and search logic moved to create contact service
  [IRISMP-113] - changes added
  Minor fixes
  [IRISMP-113] - changes added
  [IRISMP-112] - Added mp_client check on client model and used it on observers
  [IRISMP-116] - suggested changes added
  [IRISMP-112] - PR changes
  [IRISMP-116] - Send Payment link changes added
  ...

+++ app/controllers/kylas_auth_controller.rb
@@ -11,17 +11,20 @@ class KylasAuthController < ApplicationController
     if authorization_code.present?
       response = Kylas::ExchangeCode.new(authorization_code).call
       if response[:success]
-        current_user.update_tokens_details!(response)
-        current_user.update_users_and_tenants_details
-        session.delete(:previous_url) if auth_request?(session[:previous_url])
-        flash[:success] = I18n.t('kylas_auth.successfully_installed')
+        if current_user.update_users_and_tenants_details(response)
+          session.delete(:previous_url) if auth_request?(session[:previous_url])
+          flash[:success] = I18n.t('kylas_auth.successfully_installed')
+        else
+          sign_out current_user
+          redirect_to action: :authenticate, code: params[:code] and return
+        end
       else
         flash[:alert] = I18n.t('kylas_auth.facing_problem')
       end
     else
       flash[:alert] = I18n.t('kylas_auth.something_went_wrong')
     end
-    _path = reset_password_after_first_login_admin_user_path(current_user) if policy([:admin, current_user]).reset_password_after_first_login?
+    _path = reset_password_after_first_login_admin_user_path(current_user) if current_user && policy([:admin, current_user]).reset_password_after_first_login?
     redirect_to _path || root_path
   end
